Create `AbstractDatabaseCommand`
